Time stamp append hot fix
The append option for the logger was recently opened up to the rollover
logger by myself. Therefore, I'm assuming for the most part the option
is untested. I am seeing an issue with the timestamps not matching up
between two CAN buses I am logging. One of the buses has hundreds of
messages a minute, and I am therefore using `.blf`. The other bus has
about 10 messages per minute with more sensitive data, and I am
therefore using the `.log` format which contains only a single-line
buffer. The issue I am having specifically relates to the following
successive steps: 1) creation of a log file, 2) no log messages seen
on the bus and 3) power shutdown without warning (in Linux this can be
recreated with `kill <PID>`).

When creating a log file, the timestamp associated with the log creation
time is placed into the top of the file. (This cannot directly be seen
with the `.log` format, however it is still there and can be visually
inspected by converting to a `.asc` format.) When the append option is
utilized, an existing log file is opened and appended to (if any),
rather than being overwritten. This is where the timestamp issue can
be introduced. The `python-can` package will consider the timestamp of
creation of the file as the "current" timestamp. The package then
considers the "current" time at the start of the log to be when the
file was created. The rollover functionality preserves the integrity
of all instances, and therefore the timestamp issue is carried over
to each subsequent log.The nature of this bug is that the timestamp
contained in the logs that are being remotely captured appear eratic
and unpredictable.

The changes made are a modification to the startup. Rather than using
the append option, if a log file exists, then the file will be saved
as the `0th` log. Then, a new log file is created to "begin" the log.
This ensures that the `starting_timestamp` understood by the package
is the actual datetime, rather than one carried over from a previous
log.